+++ src/networkcloud/azext_networkcloud/aaz/latest/networkcloud/cluster/baremetalmachinekeyset/_update.py
@@ -13,6 +13,7 @@ from azure.cli.core.aaz import *
 
 @register_command(
     "networkcloud cluster baremetalmachinekeyset update",
+    is_preview=True,
 )
 class Update(AAZCommand):
     """Update properties of bare metal machine key set for the provided cluster, or update the tags associated with it. Properties and tag updates can be done independently.
@@ -22,9 +23,9 @@ class Update(AAZCommand):
     """
 
     _aaz_info = {
-        "version": "2025-02-01",
+        "version": "2025-07-01-preview",
         "resources": [
-            ["mgmt-plane", "/subscriptions/{}/resourcegroups/{}/providers/microsoft.networkcloud/clusters/{}/baremetalmachinekeysets/{}", "2025-02-01"],
+            ["mgmt-plane", "/subscriptions/{}/resourcegroups/{}/providers/microsoft.networkcloud/clusters/{}/baremetalmachinekeysets/{}", "2025-07-01-preview"],
         ]
     }
 
@@ -231,7 +232,7 @@ class Update(AAZCommand):
         def query_parameters(self):
             parameters = {
                 **self.serialize_query_param(
-                    "api-version", "2025-02-01",
+                    "api-version", "2025-07-01-preview",
                     required=True,
                 ),
             }
@@ -405,6 +406,9 @@ class _UpdateHelper:
         properties.privilege_level = AAZStrType(
             serialized_name="privilegeLevel",
             flags={"required": True},
+        )
+        properties.privilege_level_name = AAZStrType(
+            serialized_name="privilegeLevelName",
         )
         properties.provisioning_state = AAZStrType(
             serialized_name="provisioningState",